Moved files around to make growth easier
Firstly each function has been moved to it's own file. This will allow
simpler editing should the number of functions increase

Secondly the tests have been moved next to each file. This makes it
easier to notice when a test is missing, and allows you to find the test
quicker.
