+++ lib/f.compose.mjs
@@ -1,0 +1,7 @@
+import curry from './f.curry.mjs'
+
+const compose = (...args) =>
+  (...rest) =>
+    args.reduceRight((value, fn) => curry(fn)(value), ...rest)
+
+export default compose
