+++ lib/f.compose.test.js
