+++ lib/f.curry.mjs
@@ -1,0 +1,6 @@
+const curry = (fn, ...args) =>
+  args.length >= fn.length
+    ? fn(...args)
+    : (...rest) => curry(fn, ...args, ...rest)
+
+export default curry
