+++ lib/f.curry.test.js
