+++ lib/f.flip.test.js
