+++ lib/f.merge.mjs
@@ -1,0 +1,4 @@
+const merge = (...args) =>
+  Object.assign({}, ...args)
+
+export default merge
