+++ lib/f.merge.test.js
