+++ lib/f.mjs
@@ -1,23 +1,14 @@
-export const curry = (fn, ...args) =>
-  args.length >= fn.length
-    ? fn(...args)
-    : (...rest) => curry(fn, ...args, ...rest)
+import fCurry from './f.curry.mjs'
+import fFlip from './f.flip.mjs'
+import fCompose from './f.compose.mjs'
+import fPipe from './f.pipe.mjs'
+import fMerge from './f.merge.mjs'
 
-export const flip = (fn, ...args) =>
-  args.length >= fn.length
-    ? fn(...args.slice(0, fn.length).reverse())
-    : (...rest) => flip(fn, ...args, ...rest)
-
-export const compose = (...args) =>
-  (...rest) =>
-    args.reduceRight((value, fn) => curry(fn)(value), ...rest)
-
-export const pipe = (...args) =>
-  (...rest) =>
-    args.reduce((value, fn) => curry(fn)(value), ...rest)
-
-export const merge = (...args) =>
-  Object.assign({}, ...args)
+export const curry = fCurry
+export const flip = fFlip
+export const compose = fCompose
+export const pipe = fPipe
+export const merge = fMerge
 
 export default {
   curry,
