+++ lib/f.pipe.test.js
